random_path_generator added

+++ Logic/gimbal.py
@@ -1,7 +1,7 @@
 from servo import Servo
 
 class Gimbal:
-    def __init__(self, pinServoHorizontal: int, pinServoVertical: int):
+    def __init__(self, pinServoHorizontal: float, pinServoVertical: float):
         self.servoHorizontal = Servo(pinServoHorizontal)
         self.servoVertical = Servo(pinServoVertical)
